requirements-py2.txt
boto
dnspython
mock
psycopg2
PyYAML
requests
six >= 1.7
kazoo>=2.2.1
python-etcd>=0.4.1
click
